build: roll up the compile package type declarations into a single index.d.ts
Use `tsup` (with `dts.only`) instead of a plain `tsc --emitDeclarationOnly` run,
so that the declarations generated from the JSDoc comments are rolled up into a
single `dist/index.d.ts` containing only the public API, rather than one file per
source module.

To keep the rolled-up declarations readable, the types are now re-exported from
`index.js` with `export *` rather than being re-declared as JSDoc typedef
aliases.  Re-declaring them caused the rollup to emit a second copy of each type
with a `$1` suffix, which leaked into the generated API docs for the build
package.  For the same reason, the two `@return {ParsedModel}` tags in `index.js`
now use the explicit `import(...)` form so that they resolve to the same type.

Since `export *` would also have made `normalizeModelSpec` part of the public
API, that helper moves to its own module, leaving `_shared/model-spec.js` as a
types-only module (mirroring `_shared/model-spec.ts` in the build package).
